# ice-bubble-admin/src/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ice-bubble Admin API
  description: ice-bubble 管理后台 REST API 文档
  version: 1.4.1
  contact:
    name: ice-bubble team
servers:
  - url: http://localhost:13000
    description: 本地开发
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer Token 认证（配置在 auth.token 或 ICE_AUTH_TOKEN）
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    Session:
      type: object
      properties:
        session_key:
          type: string
        agent_id:
          type: string
          nullable: true
        agent_name:
          type: string
          nullable: true
        avatar:
          type: string
          nullable: true
        label:
          type: string
          nullable: true
        channel:
          type: string
          nullable: true
        message_count:
          type: integer
        first_message:
          type: string
          nullable: true
        last_message:
          type: string
          nullable: true
        last_message_at:
          type: string
          format: date-time
          nullable: true
        first_message_at:
          type: string
          format: date-time
          nullable: true
        session_status:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        model_provider:
          type: string
          nullable: true
        created_at:
          type: string
          nullable: true
        updated_at:
          type: string
          nullable: true
...
paths:
...
  /api/sessions:
    get:
      tags: [Sessions]
      summary: 会话列表（Admin SQLite）
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: agent_id
          in: query
          schema: { type: string }
        - name: channel
          in: query
          schema: { type: string }
        - name: platform
          in: query
          schema: { type: string }
      responses:
        '200':
          description: 会话列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  total:
                    type: integer
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Session'
...
  /api/messages:
    get:
      tags: [Messages]
      summary: 消息列表
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: session_key
          in: query
          schema: { type: string }
        - name: archived
          in: query
          description: "true 查询归档数据"
          schema: { type: string }
      responses:
        '200':
          description: 消息列表
  /api/messages/timeline:
    get:
      tags: [Messages]
      summary: 群聊风格消息时间线
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
        - name: before
          in: query
          schema: { type: string }
          description: 游标时间戳，返回此时间之前的消息
        - name: since
          in: query
          schema: { type: string }
          description: 时间戳，返回此时间之后的消息（增量轮询）
        - name: agent_ids
          in: query
          schema: { type: string }
          description: 逗号分隔的 agent_id
        - name: message_types
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: exclude_system_noise
          in: query
          schema: { type: string }
        - name: exclude_cron
          in: query
          schema: { type: string }
      responses:
        '200':
          description: 时间线数据
...
  /api/agents/overview:
    get:
      tags: [Agents]
      summary: Agent 概览（状态、当前任务、今日消息数）
      responses:
        '200':
          description: Agent 概览
        '503':
          description: 概览服务未初始化
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/agents/with-activity:
    get:
      tags: [Agents]
      summary: Agent 列表 + 活动热力图
      parameters:
        - name: days
          in: query
          schema: { type: integer, default: 90, maximum: 365 }
      responses:
        '200':
          description: Agent + 活动数据
  /api/agents/token-summary:
    get:
      tags: [Agents]
      summary: Token 统计（仅 LLM 调用消息）
      description: >
        返回每个 agent 的 token 消耗汇总。message_count 仅统计有 token 记录的消息
        （tokens_input/tokens_output/cost_total 任一非空），不含心跳、系统噪声等。
        与 /api/agents 返回的 admin_agents.message_count 不同——后者包含所有消息类型。
      parameters:
        - name: agentId
          in: query
          schema: { type: string }
        - name: date
          in: query
          schema: { type: string }
          description: YYYY-MM-DD
      responses:
        '200':
          description: Token 统计数据
...
  /api/agents/{id}/avatar:
    get:
      tags: [Agents]
      summary: 获取 Agent 头像
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 头像数据
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent_id:
                    type: string
                  avatar:
                    type: string
    put:
      tags: [Agents]
      summary: 更新 Agent 头像
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                avatar:
                  type: string
                  nullable: true
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /api/agents/{id}/activity:
    get:
      tags: [Agents]
      summary: 单 Agent 活动热力图
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: days
          in: query
          schema: { type: integer, default: 90, maximum: 365 }
      responses:
        '200':
          description: 活动数据

  # ── Modules ──
...
  /api/chat/history:
    get:
      tags: [Chat]
      summary: Gateway 聊天历史（HTTP 备用通道）
      parameters:
        - name: sessionKey
          in: query
          required: true
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: before
          in: query
          schema: { type: string }
      responses:
        '200':
          description: 聊天历史
        '502':
          description: Gateway 不可用
...